+++ src/Network/Network/Models/Cortex/PSVpnGatewayIpConfiguration.cs
@@ -14,24 +14,12 @@
 
 namespace Microsoft.Azure.Commands.Network.Models
 {
-    using System.Collections.Generic;
-    using Microsoft.WindowsAzure.Commands.Common.Attributes;
-
-    public class PSVpnGateway : PSTopLevelResource
+    public class PSVpnGatewayIpConfiguration
     {
-        public List<PSVpnConnection> Connections { get; set; }
+        public string Id { get; set; }
 
-        public PSBgpSettings BgpSettings { get; set; }
+        public string PublicIpAddress { get; set; }
 
-        [Ps1Xml(Label = "Virtual Hub", Target = ViewControl.Table, ScriptBlock = "$_.VirtualHub.Id")]
-        public PSResourceId VirtualHub { get; set; }
-
-        [Ps1Xml(Label = "Scale Unit", Target = ViewControl.Table)]
-        public int VpnGatewayScaleUnit { get; set; }
-
-        [Ps1Xml(Label = "Provisioning State", Target = ViewControl.Table)]
-        public string ProvisioningState { get; set; }
-
-        public List<PSVpnGatewayIpConfiguration> IpConfigurations { get; set;} 
+        public string PrivateIpAddress { get; set; }
     }
 }